python_api/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
dateparser